getters in ExceptionRecord, moved Transformation controller into bulkscan
