+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/bulkscan/transformation/in/ExceptionRecord.java
@@ -1,7 +1,9 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.transformation.in;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
 
+@Getter
 public class ExceptionRecord {
 
     private final String caseTypeId;
